requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
ecdsa==0.17.0
fastapi==0.75.1
greenlet==1.1.2
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
Mako==1.2.0
MarkupSafe==2.1.1
oauth2==1.9.0.post1
passlib==1.7.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.34
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.2
zipp==3.8.0
gunicorn==20.1.0
